Streamline and pythonify convert_xml_csv.ipynb
Following DRY, I moved a lot of the code to analyze a file into its own
function, and created another function to get fields out of the xmldict.
It should be a lot easier to add fields to the analysis, and the code
looks a lot cleaner now.

Also, added a basic .gitignore to ignore the data/ directory.