Speed up EIN extraction greatly
Apparently, the event loop we were running earlier:
```python
for file in files:
    one_row = extract_one(file)
    df = df.append(one_row)
```
is really slow due to the third line, since it copies the entire
DataFrame every time, resulting in quadratic behavior. To mitigate this,
I replaced it with a list comprehension like so:
```python
df = pd.concat([extract_one(file) for file in files])
```
which resulted in at least a 60x speed increase. This also means we
don't have to rely on the index files at all now!